requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-cloud-vision==3.5.0
gspread==5.11.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
requests==2.31.0
python-dotenv==1.0.0
